pyproject.toml
[tool.poetry]
name = "contest-bot"
version = "0.1.0"
description = ""
authors = ["Sanketh Varamballi <sankethv03@gmail.com>"]
readme = "README.md"
packages = [{include = "contest_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.3"
tabulate = "^0.9.0"
pony = "^0.7.16"
python-dotenv = "^1.0.0"
flashtext = "^2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
